face_cut/requirements.txt
boto3
ydb
Pillow